requirements.txt
alembic==1.13.3
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
